Split out User monkeypatching
